+++ packages/apollo-server-cloud-function/src/ApolloServer.ts
@@ -83,6 +83,12 @@ export class ApolloServer extends ApolloServerBase {
     }
 
     return (req: Request, res: Response) => {
+      // Handle both the root of the GCF endpoint and /graphql
+      if (!['', '/', '/graphql'].includes(req.path)) {
+        res.status(404).end();
+        return;
+      }
+
       if (cors) {
         if (typeof cors.origin === 'string') {
           res.set('Access-Control-Allow-Origin', cors.origin);
@@ -108,7 +114,8 @@ export class ApolloServer extends ApolloServerBase {
       }
 
       if (this.playgroundOptions && req.method === 'GET') {
-        if (req.accepts('text/html')) {
+        const acceptHeader = req.headers['accept'] as string;
+        if (acceptHeader && acceptHeader.includes('text/html')) {
           const playgroundRenderPageOptions: PlaygroundRenderPageOptions = {
             endpoint: req.get('referer'),
             ...this.playgroundOptions,