+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All of the packages in the `apollo-server` repo are released with the same versi
 - Add support for `graphql-js` v14 by augmenting typeDefs with the `@cacheControl` directive so SDL validation doesn't fail [#1595](https://github.com/apollographql/apollo-server/pull/1595)
 - Add `node-fetch` extensions typing to `RequestInit` [#1602](https://github.com/apollographql/apollo-server/pull/1602)
 - Switch `ApolloServerBase.schema` from private access to protected access. [#1610](https://github.com/apollographql/apollo-server/pull/1610)
+- Add toggle for including error messages in reports [#1615](https://github.com/apollographql/apollo-server/pull/1615)
+- Fix `apollo-server-cloud-functions` tests [#1611](https://github.com/apollographql/apollo-server/pull/1611/)
 
 ### v2.0.5